First commit of a started web app
At this moment, I have developed some of the styles, principal pages, and components of my website. We have a welcome page, an events page with filters and mocked data (using a JSON server). Additionally, there's a simulated login/registration page and a profile page. The profile page allows users to check their favorite events or the events they have created if they are event organizers.